requirements.txt
Django==4.2.5
Pillow==10.0.1
django-ckeditor==6.3.0
django-cors-headers==4.2.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
requests==2.31.0

